slice.js 파일 생성 및 reducer, actions 적용
slice.js 파일을 만들어 Redux Toolkit을 사용한다.

reducer와 actions에 있는 내용을 적용한다.
